put in guard to make sure Envfile has been sourced

+++ deploy.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ -z ${WAS_SOURCED+x} ]; then
+    echo "source Envfile"
+    exit 1
+fi
+
 rm -rf public/*
 bundle exec rake clean gen_deploy
 rm -rf docs/*
